Move tests to jest  🐿 v2.10.3
