+++ Makefile
@@ -5,7 +5,7 @@ node_modules/@financial-times/n-gage/index.mk:
 -include node_modules/@financial-times/n-gage/index.mk
 
 unit-test:
-	export PORT=5134; mocha --recursive -r loadvars.js
+	jest
 
 minus-eslint: ci-n-ui-check _verify_lintspaces _verify_pa11y_testable
