+++ lib/__mocks__/shellpromise.js
@@ -1,0 +1,2 @@
+const HEROKU_AUTH_TOKEN = 'herokuToken123';
+module.exports = () => Promise.resolve(HEROKU_AUTH_TOKEN);
