+++ lib/heroku-api.unit.test.js
@@ -1,10 +1,10 @@
+require('isomorphic-fetch');
 const nock = require('nock');
-const expect = require('chai').expect;
-const herokuApi = require('../../lib/heroku-api');
+const herokuApi = require('./heroku-api');
 
 describe('heroku-api', function () {
 	let nockScope;
-	before(() => {
+	beforeAll(() => {
 		nockScope = nock('https://api.heroku.com');
 	});
 
@@ -20,7 +20,7 @@ describe('heroku-api', function () {
 			});
 		const output = await herokuApi({ endpoint });
 
-		expect(output).to.deep.equal({
+		expect(output).toEqual({
 			something: 'yes'
 		});
 	});
@@ -37,7 +37,7 @@ describe('heroku-api', function () {
 			});
 		const { headers: { authorization } } = await herokuApi({ endpoint, authToken });
 		const [ bearerToken ] = authorization;
-		expect(bearerToken).to.equal(`Bearer ${authToken}`);
+		expect(bearerToken).toEqual(`Bearer ${authToken}`);
 	});
 
 	it('throws an error on failure', async () => {
@@ -49,8 +49,8 @@ describe('heroku-api', function () {
 		await herokuApi({ endpoint, authToken })
 			.catch((error) => {
 				const { name, status } = error;
-				expect(name).to.equal('BAD_RESPONSE');
-				expect(status).to.equal(400);
+				expect(name).toEqual('BAD_RESPONSE');
+				expect(status).toEqual(400);
 			});
 	});
 });
