+++ lib/pipelines.unit.test.js
@@ -1,6 +1,5 @@
 'use strict';
 require('isomorphic-fetch');
-const expect = require('chai').expect;
 const co = require('co');
 const mockery = require('mockery');
 const sinon = require('sinon');
