+++ lib/review-apps.js
@@ -44,7 +44,6 @@ const throwIfNotOk = async res => {
 	const { ok, status, url } = res;
 	if (!ok) {
 		const errorBody = await res.json();
-
 		console.error('Fetch error:', status, url, errorBody); // eslint-disable-line no-console
 		throw errorBody;
 	}
