+++ lib/review-apps.unit.test.js
@@ -1,12 +1,13 @@
-const expect = require('chai').expect;
+require('isomorphic-fetch');
 const nock = require('nock');
 
-const { getAppName } = require('../../lib/review-apps');
+const { getAppName } = require('./review-apps');
 
 describe('review-apps', function () {
 	let nockScope;
-	before(() => {
+	beforeAll(() => {
 		nockScope = nock('https://api.heroku.com');
+		jest.mock('shellpromise', () => a => a, { virtual: true });
 	});
 
 	afterEach(() => {
@@ -22,7 +23,7 @@ describe('review-apps', function () {
 				});
 			return getAppName(appId)
 				.then(name => {
-					expect(name).to.equal('the-app');
+					expect(name).toEqual('the-app');
 				});
 		});
 
@@ -34,7 +35,7 @@ describe('review-apps', function () {
 				});
 			return getAppName(appId)
 				.catch((error) => {
-					expect(error.message).to.equal('Some error occurred');
+					expect(error.message).toEqual('Some error occurred');
 				});
 		});
 	});
