+++ package.json
@@ -32,15 +32,11 @@
   "devDependencies": {
     "@financial-times/n-gage": "^2.0.4",
     "body-parser": "^1.14.1",
-    "chai": "^3.3.0",
     "eslint": "^2.8.0",
     "lintspaces-cli": "^0.1.1",
-    "mocha": "^5.0.0",
+    "jest": "^23.6.0",
     "mockery": "^1.4.0",
     "sinon": "^4.1.3"
-  },
-  "babel": {
-    "optional": "es7.classProperties"
   },
   "scripts": {
     "precommit": "node_modules/.bin/secret-squirrel",
